shrinkwrap: Move is-only-optional and is-only-dev out to their own files
