+++ lib/shrinkwrap.js
@@ -8,9 +8,8 @@ const readFile = BB.promisify(require('graceful-fs').readFile)
 const getRequested = require('./install/get-requested.js')
 const id = require('./install/deps.js')
 const iferr = require('iferr')
-const isDevDep = require('./install/is-dev-dep.js')
-const isOptDep = require('./install/is-opt-dep.js')
-const isProdDep = require('./install/is-prod-dep.js')
+const isOnlyOptional = require('./install/is-only-optional.js')
+const isOnlyDev = require('./install/is-only-dev.js')
 const lifecycle = require('./utils/lifecycle.js')
 const log = require('npmlog')
 const moduleName = require('./utils/module-name.js')
@@ -127,7 +126,7 @@ function shrinkwrapDeps (deps, top, tree, seen) {
       }
     }
     if (childIsOnlyDev) pkginfo.dev = true
-    if (isOptional(child)) pkginfo.optional = true
+    if (isOnlyOptional(child)) pkginfo.optional = true
     if (child.requires.length) {
       pkginfo.requires = {}
       child.requires.sort((a, b) => moduleName(a).localeCompare(moduleName(b))).forEach((required) => {
@@ -241,45 +240,3 @@ function checkPackageFile (dir, name) {
   }).catch({code: 'ENOENT'}, () => {})
 }
 
-// Returns true if the module `node` is only required direcctly as a dev
-// dependency of the top level or transitively _from_ top level dev
-// dependencies.
-// Dual mode modules (that are both dev AND prod) should return false.
-function isOnlyDev (node, seen) {
-  if (!seen) seen = {}
-  return node.requiredBy.length && node.requiredBy.every(andIsOnlyDev(moduleName(node), seen))
-}
-
-// There is a known limitation with this implementation: If a dependency is
-// ONLY required by cycles that are detached from the top level then it will
-// ultimately return true.
-//
-// This is ok though: We don't allow shrinkwraps with extraneous deps and
-// these situation is caught by the extraneous checker before we get here.
-function andIsOnlyDev (name, seen) {
-  return function (req) {
-    var isDev = isDevDep(req, name)
-    var isProd = isProdDep(req, name)
-    if (req.isTop) {
-      return isDev && !isProd
-    } else {
-      if (seen[req.path]) return true
-      seen[req.path] = true
-      return isOnlyDev(req, seen)
-    }
-  }
-}
-
-function isOptional (node, seen) {
-  if (!seen) seen = {}
-  // If a node is not required by anything, then we've reached
-  // the top level package.
-  if (seen[node.path] || node.requiredBy.length === 0) {
-    return false
-  }
-  seen[node.path] = true
-
-  return node.requiredBy.every(function (req) {
-    return isOptDep(req, node.package.name) || isOptional(req, seen)
-  })
-}
